# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Constructatech Ventures API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Service {
  id: number;
  slug: string;
  title: string;
  summary: string;
  /** @nullable */
  body?: string | null;
  includes: string[];
  icon: string;
  order: number;
}

export interface Industry {
  id: number;
  slug: string;
  name: string;
  blurb: string;
  icon: string;
  order: number;
}

export interface CompanyStats {
  projectsCompleted: number;
  clientsServed: number;
  yearsActive: number;
  industriesServed: number;
}

export interface QuoteRequestInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  company: string;
  email: string;
  phone: string;
  serviceInterest?: string[];
  /** @nullable */
  industry?: string | null;
  /** @minLength 1 */
  message: string;
  /** @nullable */
  honeypot?: string | null;
}

export type QuoteRequestStatus = typeof QuoteRequestStatus[keyof typeof QuoteRequestStatus];


export const QuoteRequestStatus = {
  new: 'new',
  contacted: 'contacted',
  quoted: 'quoted',
  won: 'won',
  lost: 'lost',
} as const;

export interface QuoteRequest {
  id: number;
  name: string;
  company: string;
  email: string;
  phone: string;
  serviceInterest?: string[];
  /** @nullable */
  industry?: string | null;
  message: string;
  status: QuoteRequestStatus;
  createdAt: string;
}

export type QuoteRequestUpdateStatus = typeof QuoteRequestUpdateStatus[keyof typeof QuoteRequestUpdateStatus];


export const QuoteRequestUpdateStatus = {
  new: 'new',
  contacted: 'contacted',
  quoted: 'quoted',
  won: 'won',
  lost: 'lost',
} as const;

export interface QuoteRequestUpdate {
  status?: QuoteRequestUpdateStatus;
}

export type LoginInputRole = typeof LoginInputRole[keyof typeof LoginInputRole];


export const LoginInputRole = {
  customer: 'customer',
  staff: 'staff',
} as const;

export interface LoginInput {
  email: string;
  password: string;
  role: LoginInputRole;
}

export interface AuthUser {
  id: number;
  email: string;
  name: string;
  role: string;
  /** @nullable */
  companyName?: string | null;
}

export interface ActivityItem {
  id: number;
  type: string;
  description: string;
  createdAt: string;
}

export interface PortalDashboard {
  activeProjects: number;
  openTickets: number;
  unpaidInvoices: number;
  totalInvoiceAmount?: number;
  recentActivity: ActivityItem[];
}

export interface Milestone {
  label: string;
  /** @nullable */
  dueDate?: string | null;
  done: boolean;
}

export type ProjectStatus = typeof ProjectStatus[keyof typeof ProjectStatus];


export const ProjectStatus = {
  scoping: 'scoping',
  'in-progress': 'in-progress',
  'on-hold': 'on-hold',
  completed: 'completed',
} as const;

export interface Project {
  id: number;
  title: string;
  description: string;
  status: ProjectStatus;
  customerId: number;
  /** @nullable */
  customerName?: string | null;
  milestones: Milestone[];
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  targetDate?: string | null;
  createdAt: string;
}

export type ProjectInputStatus = typeof ProjectInputStatus[keyof typeof ProjectInputStatus];


export const ProjectInputStatus = {
  scoping: 'scoping',
  'in-progress': 'in-progress',
  'on-hold': 'on-hold',
  completed: 'completed',
} as const;

export interface ProjectInput {
  title: string;
  description: string;
  customerId: number;
  status?: ProjectInputStatus;
  milestones?: Milestone[];
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  targetDate?: string | null;
}

export type ProjectUpdateStatus = typeof ProjectUpdateStatus[keyof typeof ProjectUpdateStatus];


export const ProjectUpdateStatus = {
  scoping: 'scoping',
  'in-progress': 'in-progress',
  'on-hold': 'on-hold',
  completed: 'completed',
} as const;

export interface ProjectUpdate {
  title?: string;
  description?: string;
  status?: ProjectUpdateStatus;
  milestones?: Milestone[];
  /** @nullable */
  targetDate?: string | null;
}

export interface TicketMessage {
  id: number;
  author: string;
  body: string;
  isStaff: boolean;
  createdAt: string;
}

export type TicketType = typeof TicketType[keyof typeof TicketType];


export const TicketType = {
  support: 'support',
  'change-request': 'change-request',
} as const;

export type TicketPriority = typeof TicketPriority[keyof typeof TicketPriority];


export const TicketPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  urgent: 'urgent',
} as const;

export type TicketStatus = typeof TicketStatus[keyof typeof TicketStatus];


export const TicketStatus = {
  open: 'open',
  'in-progress': 'in-progress',
  resolved: 'resolved',
  closed: 'closed',
} as const;

export interface Ticket {
  id: number;
  subject: string;
  type: TicketType;
  priority: TicketPriority;
  status: TicketStatus;
  customerId: number;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  projectId?: number | null;
  messages: TicketMessage[];
  createdAt: string;
}

export type TicketInputType = typeof TicketInputType[keyof typeof TicketInputType];


export const TicketInputType = {
  support: 'support',
  'change-request': 'change-request',
} as const;

export type TicketInputPriority = typeof TicketInputPriority[keyof typeof TicketInputPriority];


export const TicketInputPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  urgent: 'urgent',
} as const;

export interface TicketInput {
  /** @minLength 1 */
  subject: string;
  type: TicketInputType;
  priority: TicketInputPriority;
  /** @nullable */
  projectId?: number | null;
  /** @minLength 1 */
  body: string;
}

export interface TicketMessageInput {
  /** @minLength 1 */
  body: string;
}

export type InvoiceStatus = typeof InvoiceStatus[keyof typeof InvoiceStatus];


export const InvoiceStatus = {
  draft: 'draft',
  sent: 'sent',
  paid: 'paid',
  overdue: 'overdue',
} as const;

export interface Invoice {
  id: number;
  number: string;
  customerId: number;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  projectId?: number | null;
  amount: number;
  currency: string;
  status: InvoiceStatus;
  issueDate: string;
  dueDate: string;
  createdAt: string;
}

export type CustomerStatus = typeof CustomerStatus[keyof typeof CustomerStatus];


export const CustomerStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;

export interface Customer {
  id: number;
  companyName: string;
  contactName: string;
  email: string;
  phone: string;
  status: CustomerStatus;
  createdAt: string;
}

export interface CustomerAccount {
  id: number;
  companyName: string;
  contactName: string;
  email: string;
  phone: string;
}

export interface CustomerAccountUpdate {
  contactName?: string;
  phone?: string;
  companyName?: string;
}

export type ListPortalTicketsParams = {
status?: string;
};