# lib/api-zod/src/generated/types/quoteRequestUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Constructatech Ventures API
 * OpenAPI spec version: 0.1.0
 */
import type { QuoteRequestUpdateStatus } from './quoteRequestUpdateStatus';

export interface QuoteRequestUpdate {
  status?: QuoteRequestUpdateStatus;
}