# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Constructatech Ventures API
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: number;
  slug: string;
  title: string;
  summary: string;
  /** @nullable */
  body?: string | null;
  includes: string[];
  icon: string;
  order: number;
}